Add files via upload
For a given array with N words, find the word with length closest to the average length of the words in the array. If there are two words with lengths that are on the same distance to the average length, return the longer word. Additionally, if they are with the same length, you take the first one. For example for the array "a", "an", "cat", "door", "apple" the average word length is (1 + 2 + 3 + 4 + 5) / 5 = 15 / 5 = 3, which means that the word with length closest to the average length is "cat" (length is 3).

For the array "I", "on", "dog", "star", "water", "bright" the average word length is 3.5 and both words "dog" and "star" have lengths that are equally distant to the average length (lengths are 3 and 4, respectively). The correct answer is the longer word which is "star" (with length 4).

There can be duplicates in the array.
Input: The first number in the input is the number of words in the array N, then the words follow in each line.
Output: The found word, that is with length closest to the average word length in the array.

Pay attention:

1. All the needed code for the structure that you need to use is given. The test class ArrayMeanWordLength.java is also given, with completely implemented input and output. You only need to change the code of the String wordClosestToAverageLength(Array<String> arr) method.

2. You must not change the main method !